backend/requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
async-generator==1.10
attrs==25.4.0
base58==2.1.1
bcrypt==4.1.3
bitarray==3.8.0
black==25.12.0
blake3==1.0.8
boto3==1.42.16
botocore==1.42.16
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
ckzg==2.1.5
click==8.3.1
cryptography==46.0.3
cytoolz==1.1.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==6.0.0b1
fastapi==0.110.1
filelock==3.20.1
flake8==7.3.0
frozenlist==1.8.0
fsspec==2025.12.0
h11==0.16.0
hexbytes==1.3.1
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
ipfshttpclient==0.7.0
isort==7.0.0
Jinja2==3.1.6
jmespath==1.0.1
jq==1.10.0
librt==0.7.4
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mccabe==0.7.0
mdurl==0.1.2
mmh3==5.2.0
morphys==1.0
motor==3.3.1
mpmath==1.3.0
multiaddr==0.1.1
multidict==6.7.0
mypy==1.19.1
mypy_extensions==1.1.0
netaddr==1.3.0
networkx==3.6.1
numpy==2.4.0
oauthlib==3.3.1
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
parsimonious==0.10.0
passlib==1.7.4
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.1
pluggy==1.6.0
propcache==0.4.1
psutil==7.2.0
py-cid==0.4.0
py-multibase==2.0.0
py-multicodec==1.0.0
py-multihash==3.0.0
pyasn1==0.6.1
pycodestyle==2.14.0
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.5.0
PyPDF2==3.0.1
pytest==9.0.2
python-baseconv==1.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.21
pytokens==0.3.0
pytz==2025.2
pyunormalize==17.0.0
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rlp==4.1.0
rsa==4.9.1
s3transfer==0.16.0
s5cmd==0.2.0
safetensors==0.7.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.37.2
sympy==1.14.0
tokenizers==0.22.1
toolz==1.1.0
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.3
trio==0.32.0
trio-typing==0.10.0
typer==0.20.1
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uvicorn==0.25.0
varint==1.0.2
watchfiles==1.1.1
web3==7.14.0
websockets==15.0.1
yarl==1.22.0
zipp==3.23.0
